+++ source/ffgl/Manifest.cpp
@@ -130,6 +130,8 @@ bool Manifest::load( const std::string& path, Manifest& out, std::string& error 
 		out.supportsOpenGLRender = v->boolean();
 	if( const Json* v = root.find( "supportsMetalRender" ) )
 		out.supportsMetalRender = v->boolean();
+	if( const Json* v = root.find( "supportsOpenCLRender" ) )
+		out.supportsOpenCLRender = v->boolean();
 
 	if( out.identifier.empty() || out.bundlePath.empty() )
 	{